+++ feat/tests/test_data.py
@@ -3,12 +3,9 @@ import pandas as pd
 from pandas import DataFrame
 import numpy as np
 import os
-from feat.data import Fex, Fextractor, ImageDataset, VideoDataset
-from feat.utils.io import read_facet, read_openface, read_affectiva, get_test_data_path
+from feat.data import Fex, Fextractor
+from feat.utils.io import read_openface, get_test_data_path
 from nltools.data import Adjacency
-from feat.transforms import Rescale
-from torchvision.transforms import Compose
-from torchvision.io import read_image
 
 
 def test_info(capsys):
@@ -19,6 +16,8 @@ def test_info(capsys):
     assert importantstring in captured.out
 
 
+# TODO: write me
+@pytest.mark.skip("TODO Rewrite as we drop support for facet, affectiva")
 def test_fex():
     with pytest.raises(Exception):
         fex = Fex().read_feat()
@@ -218,6 +217,8 @@ def test_fex():
     assert n_components == stats["weights"].shape[1]
 
 
+# TODO: rewrite me
+@pytest.mark.skip("TODO Rewrite as we drop support for facet, affectiva")
 def test_fextractor():
     filename = os.path.join(get_test_data_path(), "iMotions_Test_v6.txt")
     df = read_facet(filename)
@@ -369,14 +370,6 @@ def test_openface():
     assert len(openface.calc_pspi()) == len(openface)
 
 
-def test_affectiva():
-    filename = os.path.join(
-        get_test_data_path(), "sample_affectiva-api-app_output.json"
-    )
-    affdex = Fex(read_affectiva(filename), sampling_freq=1, detector="Affectiva")
-    assert affdex.shape[1] == 33
-
-
 def test_feat():
     filename = os.path.join(get_test_data_path(), "Feat_Test.csv")
     fex = Fex(filename=filename, detector="Feat")
@@ -418,205 +411,3 @@ def test_stats():
         [np.array(range(int(len(doubled) / 2))), np.array(range(int(len(doubled) / 2)))]
     )
     assert doubled.assign(frame=frame).isc(col="AU04_r").iloc[0, 0] == 1
-
-
-def test_rescale_single_image(single_face_img):
-    img = read_image(single_face_img)
-
-    # Test Int
-    for scale in [0.5, 1.0, 2]:
-        output_size = int(img.shape[-1] * scale)
-
-        transform = Compose(
-            [Rescale(output_size, preserve_aspect_ratio=False, padding=False)]
-        )
-        transformed_img = transform(img)
-
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        transform = Compose(
-            [Rescale(output_size, preserve_aspect_ratio=True, padding=False)]
-        )
-        transformed_img = transform(img)
-
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        transform = Compose(
-            [Rescale(output_size, preserve_aspect_ratio=True, padding=True)]
-        )
-        transformed_img = transform(img)
-
-        assert transformed_img["Image"].shape[-1] == output_size
-        assert transformed_img["Image"].shape[-2] == output_size
-        assert transformed_img["Scale"] == scale
-        assert transformed_img["Padding"]["Top"] + transformed_img["Padding"][
-            "Bottom"
-        ] == (output_size - img.shape[-2] * scale)
-        assert (
-            transformed_img["Padding"]["Left"] + transformed_img["Padding"]["Right"]
-            == 0
-        )
-
-    # Test Tuple
-    for scale in [0.5, 1.0, 2]:
-        output_size = tuple((np.array(img.shape[1:]) * scale).astype(int))
-
-        transform = Compose(
-            [Rescale(output_size, preserve_aspect_ratio=False, padding=False)]
-        )
-        transformed_img = transform(img)
-
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        transform = Compose(
-            [Rescale(output_size, preserve_aspect_ratio=True, padding=False)]
-        )
-        transformed_img = transform(img)
-
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        output_size = (600 * scale, img.shape[-1] * scale)
-        transform = Compose(
-            [Rescale(output_size, preserve_aspect_ratio=True, padding=True)]
-        )
-        transformed_img = transform(img)
-        assert transformed_img["Image"].shape[1] == output_size[0]
-        assert transformed_img["Image"].shape[2] == output_size[1]
-        assert transformed_img["Scale"] == scale
-        assert transformed_img["Padding"]["Top"] + transformed_img["Padding"][
-            "Bottom"
-        ] == (600 * scale - img.shape[1] * scale)
-        assert (
-            transformed_img["Padding"]["Left"] + transformed_img["Padding"]["Right"]
-            == 0
-        )
-
-
-def test_imagedataset(single_face_img):
-    n_img = 10
-    image_file_list = [single_face_img] * n_img
-
-    img_data = ImageDataset(
-        image_file_list, output_size=None, preserve_aspect_ratio=False, padding=False
-    )
-    assert len(img_data) == n_img
-    img = img_data[0]["Image"]
-
-    # Test Int
-    for scale in [0.5, 1.0, 2]:
-        output_size = int(img.shape[-1] * scale)
-
-        img_data = ImageDataset(
-            image_file_list,
-            output_size=output_size,
-            preserve_aspect_ratio=False,
-            padding=False,
-        )
-        assert len(img_data) == n_img
-        transformed_img = img_data[0]
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        img_data = ImageDataset(
-            image_file_list,
-            output_size=output_size,
-            preserve_aspect_ratio=True,
-            padding=False,
-        )
-        assert len(img_data) == n_img
-        transformed_img = img_data[0]
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        img_data = ImageDataset(
-            image_file_list,
-            output_size=output_size,
-            preserve_aspect_ratio=False,
-            padding=True,
-        )
-        assert len(img_data) == n_img
-        transformed_img = img_data[0]
-        assert transformed_img["Image"].shape[-1] == output_size
-        assert transformed_img["Image"].shape[-2] == output_size
-        assert transformed_img["Scale"] == scale
-        assert transformed_img["Padding"]["Top"] + transformed_img["Padding"][
-            "Bottom"
-        ] == (output_size - img.shape[-2] * scale)
-        assert (
-            transformed_img["Padding"]["Left"] + transformed_img["Padding"]["Right"]
-            == 0
-        )
-
-    # Test Tuple
-    for scale in [0.5, 1.0, 2]:
-        output_size = tuple((np.array(img.shape[1:]) * scale).astype(int))
-
-        img_data = ImageDataset(
-            image_file_list,
-            output_size=output_size,
-            preserve_aspect_ratio=False,
-            padding=False,
-        )
-        assert len(img_data) == n_img
-        transformed_img = img_data[0]
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        img_data = ImageDataset(
-            image_file_list,
-            output_size=output_size,
-            preserve_aspect_ratio=True,
-            padding=False,
-        )
-        assert len(img_data) == n_img
-        transformed_img = img_data[0]
-        assert transformed_img["Image"].shape[-1] == img.shape[-1] * scale
-        assert transformed_img["Image"].shape[-2] == img.shape[-2] * scale
-        assert transformed_img["Scale"] == scale
-        for x in ["Left", "Top", "Right", "Bottom"]:
-            assert transformed_img["Padding"][x] == 0
-
-        output_size = (600 * scale, img.shape[-1] * scale)
-        img_data = ImageDataset(
-            image_file_list,
-            output_size=output_size,
-            preserve_aspect_ratio=True,
-            padding=True,
-        )
-        assert len(img_data) == n_img
-        transformed_img = img_data[0]
-        assert transformed_img["Image"].shape[1] == output_size[0]
-        assert transformed_img["Image"].shape[2] == output_size[1]
-        assert transformed_img["Scale"] == scale
-        assert transformed_img["Padding"]["Top"] + transformed_img["Padding"][
-            "Bottom"
-        ] == (600 * scale - img.shape[1] * scale)
-        assert (
-            transformed_img["Padding"]["Left"] + transformed_img["Padding"]["Right"]
-            == 0
-        )